+++ postserver/app/api/crud/room_review_dao.py
@@ -1,9 +1,15 @@
+from app.api.crud.room_dao import RoomDAO
 from app.model.room_review import RoomReview
+from app.errors.http_error import NotFoundError
+from app.errors.bookbnb_error import NoRelationError
 
 class RoomReviewDAO:
 
 	@classmethod
 	def add_new_room_review(cls, db, room_id, room_review_args):
+		if not RoomDAO.room_is_present(db, room_id):
+			raise NotFoundError('room')
+
 		new_room_review = RoomReview(review=room_review_args.review,
 									 room_id=room_id,
 			                         reviewer=room_review_args.reviewer,
@@ -28,34 +34,51 @@ class RoomReviewDAO:
 
 	@classmethod
 	def get_room_review(cls, db, room_id, review_id):
+		if not RoomDAO.room_is_present(db, room_id):
+			raise NotFoundError('room')
+
 		room_review = db.query(RoomReview).get(review_id)
 
-		# this should return an error in case of  
-		# the rating do not for the specified room
+		if room_review is None:
+			raise NotFoundError('room review')
+
+		if not room_review.is_from(room_id):
+			raise NoRelationError('room', 'room review')
 
 		return room_review.serialize()
 
 
 	@classmethod
 	def delete_room_review(cls, db, room_id, review_id):
+		if not RoomDAO.room_is_present(db, room_id):
+			raise NotFoundError('room')
+
 		room_review = db.query(RoomReview).get(review_id)
 		
+		if room_review is None:
+			raise NotFoundError('room review')
+
+		if not room_review.is_from(room_id):
+			raise NoRelationError('room', 'room review')	
+
 		db.delete(room_review)
 		db.commit()
-
-		# this should return an error in case of  
-		# the rating do not for the specified room
 
 		return room_review.serialize()
 
 
 	@classmethod
 	def update_room_review(cls, db, room_id, review_id, update_args):
+		if not RoomDAO.room_is_present(db, room_id):
+			raise NotFoundError('room')
+
 		room_review = db.query(RoomReview).get(review_id)
 
-		# we should see if is necessary to update
-		# owner and owner id. May be this should
-		# be a restricted method
+		if room_review is None:
+			raise NotFoundError('room review')
+
+		if not room_review.is_from(room_id):
+			raise NoRelationError('room', 'room review')
 
 		if update_args.review is not None:
 			room_review.review = update_args.review
